+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ========================
 
 * [OLMIS-2837](https://openlmis.atlassian.net/browse/OLMIS-2837): Fixed filtering on the manage POD page.
+* [OLMIS-2724](https://openlmis.atlassian.net/browse/OLMIS-2724): Fixed broken requesting facility filter select on Order View.
 
 5.0.2 / 2017-06-22
 ==================